lab1/Validation.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace lab1
{
    public class Validation
    {
        public bool IsValidNumber(string input)
        {
            foreach (char c in input)
            {
                if (!char.IsDigit(c))
                {
                    return false;
                }
            }

            if (string.IsNullOrEmpty(input))
            {
                return false;
            }

            if (input.Length > 3)
            {
                return false;
            }

            if (Convert.ToInt32(input) <= 0)
            {
                return false;
            }

            return true;
        }
    }
}
